@@ -17,7 +17,7 @@
     "electron:build": "build -c.extraMetadata.main=build/electron.js -mwl -p never"
   },
   "lint-staged": {
-    "*.{ts,js,json}": [
+    "*.{ts,js,tsx}": [
       "prettier --write",
       "git add"
     ]
@@ -38,7 +38,9 @@
   },
   "dependencies": {
     "fuse.js": "^3.2.0",
+    "humps": "^2.0.1",
     "localforage": "^1.7.1",
+    "normalizr": "^3.2.4",
     "react": "^16.3.2",
     "react-dom": "^16.3.2",
     "react-redux": "^5.0.7",
@@ -46,11 +48,13 @@
     "react-router-config": "^1.0.0-beta.4",
     "react-router-dom": "^4.2.2",
     "react-select": "^1.2.1",
-    "redux-persist": "^5.9.1"
+    "redux-persist": "^5.9.1",
+    "redux-saga": "^0.16.0"
   },
   "devDependencies": {
+    "@types/humps": "^1.1.2",
     "@types/react": "^16.3.14",
-    "@types/react-beautiful-dnd": "^6.0.1",
+    "@types/react-beautiful-dnd": "^7.1.0",
     "@types/react-dom": "^16.0.5",
     "@types/react-redux": "^6.0.0",
     "@types/react-relay": "^1.3.5",
@@ -58,25 +62,26 @@
     "@types/react-router-dom": "^4.2.6",
     "@types/react-select": "^1.2.6",
     "awesome-typescript-loader": "^5.0.0",
-    "babel-eslint": "^8.2.3",
+    "babel-eslint": "^8.2.4",
     "concurrently": "^3.5.1",
     "copy-webpack-plugin": "^4.5.1",
     "css-loader": "^0.28.11",
-    "dotenv-webpack": "^1.5.5",
     "electron": "^2.0.2",
     "electron-builder": "^20.15.1",
+    "electron-debug": "^2.0.0",
     "electron-devtools-installer": "^2.2.4",
-    "eslint": "^4.19.1",
+    "eslint": "^5.0.0",
     "eslint-plugin-react": "^7.8.2",
     "file-loader": "^1.1.11",
     "html-webpack-plugin": "^3.2.0",
     "lint-staged": "^7.1.0",
     "node-sass": "^4.9.0",
     "sass-loader": "^7.0.1",
+    "source-map-loader": "^0.2.3",
     "style-loader": "^0.21.0",
     "typescript": "^2.8.3",
     "webpack": "^4.8.3",
-    "webpack-cli": "^2.1.3",
+    "webpack-cli": "^3.0.8",
     "webpack-serve": "^1.0.2"
   }
 }
