@@ -58,7 +58,7 @@ const createWindow = () => {
 
   if (process.platform === 'darwin')
     template.push({
-      label: 'Foxtrot',
+      label: 'Trax',
       submenu: [
         { role: 'about' },
         { role: 'hide' },
@@ -78,14 +78,14 @@ const createWindow = () => {
     width: 900,
     height: 680,
     webPreferences: {
-      // nodeIntegration: false
-      webSecurity: false,
-    },
+      nodeIntegration: false,
+      preload: path.join(__dirname, 'preload.js')
+    }
   })
 
   mainWindow.loadURL(
     isDev
-      ? 'http://localhost:3000'
+      ? 'https://localhost:3000'
       : `file://${path.join(__dirname, '../build/index.html')}`
   )
   mainWindow.maximize()
