@@ -1,7 +1,8 @@
 import { all } from 'redux-saga/effects'
 
 import alertSaga from 'sagas/alertSaga'
+import userSaga from 'sagas/userSaga'
 
 export default function* rootSaga() {
-  yield all([alertSaga()])
+  yield all([alertSaga(), userSaga()])
 }
