@@ -9,16 +9,18 @@ const ExternalLinkIcon: React.SFC<TIconProps> = (props) => {
   const { id, viewbox } = props
 
   return (
-    <svg id={id} version="1.1" viewBox={viewbox} xmlns="http://www.w3.org/2000/svg">
-      <path
-        d="M12.801,66.102 L48.801,66.102 C55.4026,66.102 60.801,60.7036 60.801,54.102 L60.801,32.102 C60.801,31.0004 59.90256,30.102 58.801,30.102 C57.6994,30.102 56.801,31.00044 56.801,32.102 L56.801,54.102 C56.801,58.5004 53.1994,62.102 48.801,62.102 L12.801,62.102 C8.4026,62.102 4.801,58.5004 4.801,54.102 L4.801,18.102 C4.801,13.7036 8.4026,10.102 12.801,10.102 L34.801,10.102 C35.9026,10.102 36.801,9.20356 36.801,8.102 C36.801,7.0004 35.90256,6.102 34.801,6.102 L12.801,6.102 C6.1994,6.102 0.801,11.5004 0.801,18.102 L0.801,54.102 C0.801,60.6997 6.1994,66.102 12.801,66.102 Z"
-        id="Shape"
-      />
-      <path
-        d="M67.199,20.199 L67.199,1.898 L67.199,1.49956 L67.199,1.398 C67.199,1.29644 67.199,1.19878 67.09744,1.09722 L66.999784,0.999564 C66.999784,0.898004 66.898224,0.800344 66.898224,0.800344 C66.796664,0.601124 66.597444,0.401904 66.296664,0.198784 C66.195104,0.097224 66.097444,0.097224 65.995884,0.097224 L65.99979,0.10113 L65.69901,-0.00043 L65.59745,-0.00043 L65.19901,-0.00043 L46.89801,-0.00043 C45.79641,-0.00043 44.89801,0.89801 44.89801,1.99957 C44.89801,3.10113 45.79645,3.99957 46.89801,3.99957 L60.39801,3.99957 L31.10101,33.10157 C30.30023,33.90235 30.30023,35.10157 31.10101,35.90237 C31.90179,36.70315 33.10101,36.70315 33.90181,35.90237 L63.20281,6.70337 L63.20281,20.20337 C63.20281,21.30497 64.10125,22.20337 65.20281,22.20337 C66.30051,22.199464 67.19891,21.30103 67.19891,20.19947 L67.199,20.199 Z"
-        id="Shape"
-      />
-    </svg>
+    <span className="icon">
+      <svg id={id} version="1.1" viewBox={viewbox} xmlns="http://www.w3.org/2000/svg">
+        <path
+          d="M12.801,66.102 L48.801,66.102 C55.4026,66.102 60.801,60.7036 60.801,54.102 L60.801,32.102 C60.801,31.0004 59.90256,30.102 58.801,30.102 C57.6994,30.102 56.801,31.00044 56.801,32.102 L56.801,54.102 C56.801,58.5004 53.1994,62.102 48.801,62.102 L12.801,62.102 C8.4026,62.102 4.801,58.5004 4.801,54.102 L4.801,18.102 C4.801,13.7036 8.4026,10.102 12.801,10.102 L34.801,10.102 C35.9026,10.102 36.801,9.20356 36.801,8.102 C36.801,7.0004 35.90256,6.102 34.801,6.102 L12.801,6.102 C6.1994,6.102 0.801,11.5004 0.801,18.102 L0.801,54.102 C0.801,60.6997 6.1994,66.102 12.801,66.102 Z"
+          id="Shape"
+        />
+        <path
+          d="M67.199,20.199 L67.199,1.898 L67.199,1.49956 L67.199,1.398 C67.199,1.29644 67.199,1.19878 67.09744,1.09722 L66.999784,0.999564 C66.999784,0.898004 66.898224,0.800344 66.898224,0.800344 C66.796664,0.601124 66.597444,0.401904 66.296664,0.198784 C66.195104,0.097224 66.097444,0.097224 65.995884,0.097224 L65.99979,0.10113 L65.69901,-0.00043 L65.59745,-0.00043 L65.19901,-0.00043 L46.89801,-0.00043 C45.79641,-0.00043 44.89801,0.89801 44.89801,1.99957 C44.89801,3.10113 45.79645,3.99957 46.89801,3.99957 L60.39801,3.99957 L31.10101,33.10157 C30.30023,33.90235 30.30023,35.10157 31.10101,35.90237 C31.90179,36.70315 33.10101,36.70315 33.90181,35.90237 L63.20281,6.70337 L63.20281,20.20337 C63.20281,21.30497 64.10125,22.20337 65.20281,22.20337 C66.30051,22.199464 67.19891,21.30103 67.19891,20.19947 L67.199,20.199 Z"
+          id="Shape"
+        />
+      </svg>
+    </span>
   )
 }
 
